+++ src/main/java/com/sports/server/auth/application/AuthService.java
@@ -1,6 +1,6 @@
 package com.sports.server.auth.application;
 
-import com.sports.server.auth.JwtProvider;
+import com.sports.server.auth.jwt.JwtProvider;
 import com.sports.server.auth.dto.JwtResponse;
 import com.sports.server.auth.dto.LoginVO;
 import com.sports.server.auth.exception.AuthorizationErrorMessages;
